ref: Code refactor
