+++ packages/account/src/Components/trading-assessment/index.js
@@ -4,4 +4,3 @@ export { default as RiskToleranceWarningModal } from './risk-tolerance-warning-m
 export { default as CoolDownWarningModal } from './cooldown-warning-modal.jsx';
 export { default as VerifiedAccountModal } from './verified-account-modal.jsx';
 export { default as TestWarningModal } from './test-warning-modal.jsx';
-export { default as TradingExperienceModal } from './trading-experience-modal.jsx';
