+++ packages/account/src/Components/trading-assessment/trading-assessment-existing-user.jsx
@@ -1,141 +1,15 @@
 import React from 'react';
-import { Formik, Form } from 'formik';
-import { Text, Button, Modal, DesktopWrapper, MobileDialog, MobileWrapper } from '@deriv/components';
-import { getDefaultFields, isMobile } from '@deriv/shared';
-import { localize, Localize } from '@deriv/translations';
+import { Modal, DesktopWrapper, MobileDialog, MobileWrapper } from '@deriv/components';
+import { getDefaultFields } from '@deriv/shared';
+import { localize } from '@deriv/translations';
 import { connect } from 'Stores/connect';
 import { trading_assessment_form_config, trading_assessment_questions } from 'Configs/trading-assessment-config';
-import TradingAssessmentRadioOption from './trading-assessment-radio-buttons';
-import TradingAssessmentDropdownOption from './trading-assessment-dropdown';
-
-export const TradingAssessmentForm = ({ assessment_questions, value, onSubmit }) => {
-    const [is_next_button_enabled, setIsNextButtonEnabled] = React.useState(false);
-    const [current_question_details, setCurrentQuestion] = React.useState({
-        current_question_index: 0,
-        current_question: {},
-    });
-
-    const last_question_index = assessment_questions.length - 1; // eslint-disable-line no-unused-vars
-
-    React.useEffect(() => {
-        setCurrentQuestion(prevState => ({
-            ...prevState,
-            current_question: assessment_questions[prevState.current_question_index],
-        }));
-    }, []);
-    // eslint-disable-next-line no-unused-vars
-    const handleNextButton = () => {
-        const next_question = current_question_details.current_question_index + 1;
-        if (next_question < assessment_questions.length) {
-            setCurrentQuestion({
-                current_question_index: next_question,
-                current_question: assessment_questions[next_question],
-            });
-        }
-    };
-
-    const handlePrevButton = () => {
-        const prev_question = current_question_details.current_question_index - 1;
-        if (prev_question >= 0) {
-            setCurrentQuestion({
-                current_question_index: prev_question,
-                current_question: assessment_questions[prev_question],
-            });
-        }
-    };
-
-    const handleValueSelection = (e, form_control, callBackFn, values) => {
-        if (typeof e.persist === 'function') e.persist();
-        callBackFn(form_control, e.target.value);
-        const latest_value = { ...values, [form_control]: e.target.value };
-        if (latest_value.risk_tolerance === 'No') {
-            onSubmit(latest_value);
-        }
-    };
-
-    const isAssessmentCompleted = answers => Object.values(answers).every(answer => Boolean(answer));
-    return (
-        <div className='trading-assessment'>
-            <Text as='p' color='prominent' size='xxs' className='trading-assessment__side-note'>
-                <Localize i18n_default_text='In providing our services to you, we are required to obtain information from you in order to asses whether a given product or service is appropriate for you.' />
-            </Text>
-            <section className='trading-assessment__header'>
-                <div className='trading-assessment__header--background'>
-                    <Text as='h1' color='prominent' weight='bold' size='xs'>
-                        {current_question_details.current_question_index + 1} {localize('of')}{' '}
-                        {assessment_questions.length}
-                    </Text>
-                </div>
-                <div className='trading-assessment__header--arrow-down' />
-            </section>
-            <section className='trading-assessment__form'>
-                <Formik initialValues={{ ...value }} onSubmit={onSubmit}>
-                    {({ setFieldValue, values }) => {
-                        const { question_text, form_control, answer_options, questions } =
-                            current_question_details.current_question;
-
-                        return (
-                            <Form className='trading-assessment__form--layout'>
-                                <div className='trading-assessment__form--fields'>
-                                    {questions && questions.length ? (
-                                        <TradingAssessmentDropdownOption
-                                            item_list={questions}
-                                            onChange={handleValueSelection}
-                                            values={values}
-                                            setFieldValue={setFieldValue}
-                                            setEnableNextSection={setIsNextButtonEnabled}
-                                        />
-                                    ) : (
-                                        <TradingAssessmentRadioOption
-                                            text={question_text}
-                                            list={answer_options ?? []}
-                                            onChange={e => handleValueSelection(e, form_control, setFieldValue, values)}
-                                            values={values}
-                                            form_control={form_control}
-                                            setEnableNextSection={setIsNextButtonEnabled}
-                                        />
-                                    )}
-                                </div>
-                                <Modal.Footer
-                                    has_separator
-                                    is_bypassed={isMobile()}
-                                    className='trading-assessment__existing_btn '
-                                >
-                                    <Button.Group>
-                                        <Button
-                                            has_effect
-                                            onClick={handlePrevButton}
-                                            text={localize('Previous')}
-                                            type='button'
-                                            secondary
-                                            large
-                                        />
-                                        <Button
-                                            has_effect
-                                            is_disabled={!is_next_button_enabled}
-                                            onClick={!isAssessmentCompleted && handleNextButton}
-                                            type={isAssessmentCompleted ? 'submit' : 'button'}
-                                            text={localize('Next')}
-                                            large
-                                            primary
-                                        />
-                                    </Button.Group>
-                                </Modal.Footer>
-                            </Form>
-                        );
-                    }}
-                </Formik>
-            </section>
-        </div>
-    );
-};
+import TradingAssessmentForm from './trading-assessment-form';
 
 const TradingAssessmentExistingUser = ({
     real_account_signup_target,
     is_trade_assessment_incomplete,
     realAccountSignup,
-    setShouldShowRiskToleranceWarningModal,
-    should_show_risk_tolerance_warning_modal,
 }) => {
     const [form_values, setStateItems] = React.useState({});
 
@@ -160,7 +34,7 @@ const TradingAssessmentExistingUser = ({
                 <Modal is_open={is_trade_assessment_incomplete} title={localize('Trading Experience Assessment')}>
                     <TradingAssessmentForm
                         assessment_questions={trading_assessment_questions}
-                        value={form_values}
+                        form_values={form_values}
                         onSubmit={handleSubmit}
                     />
                 </Modal>
@@ -185,6 +59,4 @@ export default connect(({ client, ui }) => ({
     is_trade_assessment_incomplete: client.is_trading_experience_incomplete,
     real_account_signup_target: ui.real_account_signup_target,
     realAccountSignup: client.realAccountSignup,
-    should_show_risk_tolerance_warning_modal: ui.should_show_risk_tolerance_warning_modal,
-    setShouldShowRiskToleranceWarningModal: ui.setShouldShowRiskToleranceWarningModal,
 }))(TradingAssessmentExistingUser);
